@@ -7,20 +7,34 @@
   "main": "./index.js",
   "babel": {
     "presets": [
-      "es2015",
-      "react"
+      [
+        "env",
+        {
+          "targets": {
+            "browsers": [
+              "last 2 versions",
+              "safari >= 7"
+            ]
+          }
+        }
+      ],
+      [
+        "react"
+      ]
     ]
   },
   "dependencies": {
+    "babel-loader": "^7.1.2",
     "bootstrap": "^3.3.7",
     "bootswatch": "^3.3.7",
     "browser-request": "^0.3.3",
     "css-loader": "^0.28.7",
     "deep-equal": "^1.0.1",
     "deepcolour": "^1.3.0",
+    "extract-text-webpack-plugin": "^3.0.2",
     "fetch": "^1.1.0",
     "file-loader": "^0.11.1",
-    "google-fonts-webpack-plugin": "^0.3.1",
+    "google-fonts-webpack-plugin": "^0.4.4",
     "highlight.js": "^9.5.0",
     "horten": "^0.8.3",
     "inherits": "^2.0.1",
@@ -28,31 +42,28 @@
     "less": "^2.7.2",
     "less-loader": "^4.0.3",
     "lodash": "^4.17.4",
-    "react": "^15.3.0",
-    "react-bootstrap": "^0.30.2",
-    "react-dom": "^15.3.2",
-    "react-markdown": "^2.5.0",
-    "react-syntax-highlighter": "^5.7.0",
-    "react-textarea-autosize": "^5.1.0",
+    "react": "^16.2.0",
+    "react-bootstrap": "^0.32.1",
+    "react-dom": "^16.2.0",
+    "react-markdown": "^3.1.5",
+    "react-syntax-highlighter": "^7.0.0",
+    "react-textarea-autosize": "^5.2.1",
     "request": "^2.82.0",
     "request-promise": "^4.2.2",
-    "string2png": "^0.5.2",
+    "string2png": "^0.5.3",
     "style-loader": "^0.16.1",
     "url-loader": "^0.5.8",
     "webpack-dev-server": "^2.8.2"
   },
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-env": "^1.6.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
     "babelify": "^7.0.2",
-    "bootstrap": "^3.3.7",
     "browserify": "^12.0.1",
     "chai": "^3.5.0",
     "copy-webpack-plugin": "^4.0.1",
     "del": "^1.1.1",
+    "html-webpack-plugin": "^2.30.1",
     "webpack": "^2.4.1"
   },
   "engines": {
