@@ -1,6 +1,6 @@
 'use strict'
 module.exports = HortenControl
-require('./index.less')
+
 
 const _ = require('lodash')
     , React = require('react')
@@ -37,10 +37,14 @@ function HortenControl( props ) {
 
   let _class = props._class
 
-  if ( !_class )
+  if ( !_class ) {
+    console.warn('meta dump', props.meta )
     return (
       <div className="horten error">No class found for type { type }!</div>
     )
+  }
+
+  console.info('meta dump', props.path, meta.get('type/angle') )
 
   return (
     <_class
